compra articulo genera solicitud si es a fabrica

+++ Core/Core.Dominio/AggregatesModel/Compra/CompraArticuloDetalle.cs
@@ -17,7 +17,7 @@ namespace Core.Dominio.AggregatesModel
         [Column(TypeName = "decimal(18,2)")]        
         public decimal Monto { get; set; }
         [Column(TypeName = "decimal(18,2)")]
-        public double Cantidad { get; set; }
+        public int Cantidad { get; set; }
         public string Comentario { get; set; }
